Split PVec/NDVec in own files.
